Migrated tests folder, corrected some pathing
