+++ python/lolopy/__init__.py
@@ -1,1 +1,1 @@
-from .__version__ import __version__  # noqa: F401
+from .version import __version__  # noqa: F401
