+++ python/requirements.txt
@@ -1,2 +1,3 @@
+numpy
 scikit-learn==1.6.1
 py4j==0.10.9.9
