+++ python/setup.py
@@ -4,7 +4,7 @@ from shutil import copy, rmtree
 
 # single source of truth for package version
 this_directory = Path(__file__).parent.absolute()
-version_file = this_directory / "lolopy" / "__version__.py"
+version_file = this_directory / "lolopy" / "version.py"
 readme_file = this_directory / "README.md"
 
 about = {}
@@ -35,7 +35,7 @@ copy(JAR_FILE[0], jar_path / "lolo-jar-with-dependencies.jar")
 setup(
     name="lolopy",
     version=about["__version__"],
-    python_requires='>=3.9',
+    python_requires='>=3.8',
     url="https://github.com/CitrineInformatics/lolo",
     maintainer="Maxwell Venetos",
     maintainer_email="mvenetos@citrine.io",
@@ -46,6 +46,7 @@ setup(
     include_package_data=True,
     package_data={"lolopy.jar": ["*.jar"]},
     install_requires=[
+        "numpy",
         "scikit-learn<1.7",
         "py4j<0.10.10"
     ],
@@ -54,6 +55,7 @@ setup(
     long_description_content_type="text/markdown",
     classifiers=[
         'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
